requirements.txt
pandas
beautifulsoup4
requests
selenium
psycopg2
python_dotenv
sqlalchemy